feat: Create API Model
In this commit, I have created the API Model to represent the data received from the simulated product API. The API Model is responsible for structuring and mapping the data from the external source.

Changes made:
- Created the API Model to represent product data
- Defined the necessary properties to store product information
- Implemented the mapping of the received data to the API Model

This commit is an important step towards integrating with the simulated product API as it establishes the necessary data structure to correctly receive and handle the information.

Signed-off-by: ericksens <ericksens_87@hotmail.com>

#456